update can now do directories

+++ update.sh
@@ -5,8 +5,14 @@
 if [[  $# -ne 2 || $2 -le 0 ]]; then
 	echo "Incorrect input: Supply the file to transfer and the number of VM instances"
 else
-	for i in `seq 1 $2`; do
-		gcloud compute scp $1 earth-$i:~/ --ssh-key-file=~/.ssh/google-cloud-cs123
-	done
+	if [[ -f $1 ]]; then
+		for i in `seq 1 $2`; do
+			gcloud compute scp $1 earth-$i:~/ --ssh-key-file=~/.ssh/google-cloud-cs123
+		done
+	else
+		for i in `seq 1 $2`; do
+			gcloud compute scp --recurse $l earth-$i:~/ --ssh-key-file=~/.ssh/google-cloud-cs123
+		done
+	fi
 fi
